Fix #665 	Optimization of tags handling for huge repositories

+++ community/logicaldoc/trunk/logicaldoc-webservice/src/main/java/com/logicaldoc/webservice/search/WSTagCloud.java
@@ -12,7 +12,7 @@ public class WSTagCloud {
 
 	private String tag;
 
-	private int count;
+	private long count;
 
 	private int scale;
 
@@ -20,7 +20,7 @@ public class WSTagCloud {
 		this.tag = tag;
 	}
 
-	public WSTagCloud(String tag, int count) {
+	public WSTagCloud(String tag, long count) {
 		this.tag = tag;
 		this.count = count;
 	}
@@ -36,11 +36,11 @@ public class WSTagCloud {
 		this.tag = tag;
 	}
 
-	public int getCount() {
+	public long getCount() {
 		return count;
 	}
 
-	public void setCount(int count) {
+	public void setCount(long count) {
 		this.count = count;
 	}
